// src/components/ConfiguracoesFiscais/types.ts
export interface ConfiguracoesFiscaisFormData {
  certificado_a1_data: string;
  certificado_a1_senha: string;
  serie_nfe: number;
  serie_nfce: number;
  csc_id: string;
  csc_token: string;
  email_padrao_envio: string;
  layout_danfe: string;
  proxima_num_nf: number;
  regime_fiscal: string;
  regime_tributario: string;
  enviar_nfe_por_email: boolean;
}